backend/requirements.txt
Django==4.1.5
psycopg2==2.9.3
djangorestframework==3.13.0
